add "--param-type" option to ros2param list.
