+++ ros2param/ros2param/api/__init__.py
@@ -185,6 +185,22 @@ def call_list_parameters(*, node, node_name, prefix=None):
     return response.result.names
 
 
+def get_parameter_type_string(parameter_type):
+    mapping = {
+        ParameterType.PARAMETER_BOOL: 'boolean',
+        ParameterType.PARAMETER_INTEGER: 'integer',
+        ParameterType.PARAMETER_DOUBLE: 'double',
+        ParameterType.PARAMETER_STRING: 'string',
+        ParameterType.PARAMETER_BYTE_ARRAY: 'byte array',
+        ParameterType.PARAMETER_BOOL_ARRAY: 'boolean array',
+        ParameterType.PARAMETER_INTEGER_ARRAY: 'integer array',
+        ParameterType.PARAMETER_DOUBLE_ARRAY: 'double array',
+        ParameterType.PARAMETER_STRING_ARRAY: 'string array',
+        ParameterType.PARAMETER_NOT_SET: 'not set',
+    }
+    return mapping[parameter_type]
+
+
 class ParameterNameCompleter:
     """Callable returning a list of parameter names."""
